requirements.txt
# Core Framework
fastapi==0.115.6
uvicorn==0.34.0

# Database
sqlalchemy==2.0.32
asyncpg==0.30.0
psycopg2-binary==2.9.10

# Authentication & Security
python-jose==3.5.0
passlib==1.7.4
bcrypt==4.0.1
python-multipart==0.0.21

# Environment Variables
python-dotenv==1.0.1

# Validation
pydantic==2.12.5
pydantic-settings==2.12.0
email-validator==2.3.0

# Email
fastapi-mail==1.4.1

# Background Jobs
celery==5.3.4
redis==5.0.1
flower==2.0.1
